refactor: migrate all imports to path aliases
