+++ src/components/index.ts
@@ -1,12 +1,14 @@
-export { default as BackToMainButton } from './BackToMainButton';
-export { default as ErrorBoundary } from './ErrorBoundary';
-export { default as Footer } from './Footer';
-export { default as Loader } from './Loader';
-export { default as Pagination } from './Pagination';
-export { default as Pokemon } from './pokemons/Pokemon';
-export { default as PokemonDetails } from './pokemons/PokemonDetails';
-export { default as PokemonList } from './pokemons/PokemonList';
-export { default as SearchPanel } from './search/SearchPanel';
-export { default as SearchForm } from './search/SearchForm';
-export { default as SearchResultsHeader } from './search/SearchResultsHeader';
-export { default as ErrorMessage } from './ErrorMessage';
+export { default as BackToMainButton } from '@/components/BackToMainButton';
+export { default as ErrorBoundary } from '@/components/ErrorBoundary';
+export { default as ErrorMessage } from '@/components/ErrorMessage';
+export { default as Footer } from '@/components/Footer';
+export { default as Loader } from '@/components/Loader';
+export { default as Pagination } from '@/components/Pagination';
+
+export { default as Pokemon } from '@/components/pokemons/Pokemon';
+export { default as PokemonDetails } from '@/components/pokemons/PokemonDetails';
+export { default as PokemonList } from '@/components/pokemons/PokemonList';
+
+export { default as SearchForm } from '@/components/search/SearchForm';
+export { default as SearchPanel } from '@/components/search/SearchPanel';
+export { default as SearchResultsHeader } from '@/components/search/SearchResultsHeader';
