+++ src/components/pokemons/PokemonDetails.tsx
@@ -1,9 +1,9 @@
 import { useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 
-import { BackToMainButton, Loader } from '..';
-import { useGetPokemonQuery } from '../../shared/api/pokemonApiSlice';
-import { btnPrimaryStyle } from '../../shared/constants';
+import { useGetPokemonQuery } from '@/shared/api/pokemonApiSlice';
+import { btnPrimaryStyle } from '@/shared/constants';
+import { BackToMainButton, Loader } from '@components';
 
 const PokemonDetails = () => {
   const { pokemonName } = useParams();
