+++ src/main.tsx
@@ -2,10 +2,9 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 
-import './index.css';
-
-import App from './App.tsx';
-import store from './shared/store';
+import App from '@/App';
+import '@/index.css';
+import store from '@/shared/store';
 
 let root = document.getElementById('root');
