+++ src/pages/Page404.tsx
@@ -1,5 +1,5 @@
-import page404 from '../assets/page404.png';
-import { BackToMainButton } from '../components';
+import page404 from '@/assets/page404.png';
+import { BackToMainButton } from '@components';
 
 const Page404 = () => {
   return (
